refactor(signalr): remove SignalR server & `msgpack_lazer_api`
Maybe we can make `msgpack_lazer_api` independent?
