--- app/signalr/__init__.py
@@ -1,5 +1,0 @@
-from __future__ import annotations
-
-from .router import router as signalr_router
-
-__all__ = ["signalr_router"]
