--- packages/msgpack_lazer_api/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "msgpack-lazer-api"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[lib]
-name = "msgpack_lazer_api"
-crate-type = ["cdylib"]
-
-[dependencies]
-chrono = "0.4.41"
-pyo3 = { version = "0.25.0", features = ["extension-module", "chrono"] }
-rmp = "0.8.14"
