--- packages/msgpack_lazer_api/msgpack_lazer_api.pyi
@@ -1,4 +1,0 @@
-from typing import Any
-
-def encode(obj: Any) -> bytes: ...
-def decode(data: bytes) -> Any: ...
