--- packages/msgpack_lazer_api/pyproject.toml
@@ -1,16 +1,0 @@
-[build-system]
-requires = ["maturin>=1.9,<2.0"]
-build-backend = "maturin"
-
-[project]
-name = "msgpack-lazer-api"
-requires-python = ">=3.12"
-classifiers = [
-    "Programming Language :: Rust",
-    "Programming Language :: Python :: Implementation :: CPython",
-    "Programming Language :: Python :: Implementation :: PyPy",
-]
-dynamic = ["version"]
-
-[tool.maturin]
-features = ["pyo3/extension-module"]
